// apps/desktop/src/renderer/game/office-object-catalog.generated.ts
export interface OfficeObjectVariant {
  readonly scale: 16 | 32 | 48;
  readonly path: string;
  readonly w: number;
  readonly h: number;
  readonly bounds: {
    readonly x: number;
    readonly y: number;
    readonly w: number;
    readonly h: number;
  };
}