// apps/desktop/src/renderer/game/room-builder-tiles.generated.ts
export interface RoomBuilderTile {
  readonly id: string;
  readonly col: number;
  readonly row: number;
  readonly path: string;
}